[feature-102/adding_room_orm] Refactor on imports
