+++ app/src/main/java/com/manday/marvel/data/datasource/net/retrofit/RetrofitDataSource.kt
@@ -1,12 +1,14 @@
-package com.manday.marvel.data.datasource.net
+package com.manday.marvel.data.datasource.net.retrofit
 
 
-import com.manday.marvel.data.models.*
 import com.manday.core_data.utils.BASE_URL
 import com.manday.core_data.RetrofitController
 import com.manday.core_data.RetrofitController.unwrapResponse
-import com.manday.core_data.RetrofitController.unwrapResponseSingle
 import com.manday.marvel.BuildConfig
+import com.manday.marvel.data.datasource.NetDataSource
+import com.manday.marvel.data.datasource.net.api.CharacterApi
+import com.manday.marvel.data.datasource.net.models.CharacterEntity
+import com.manday.marvel.data.datasource.net.models.toCharacterEntityList
 import javax.inject.Inject
